attack_experiments/requirements.txt
numpy
scipy
pandas
scikit-learn
matplotlib
seaborn

torch
torchvision
torchaudio
torchinfo

opencv-python
Pillow
scikit-image

tqdm
pyyaml
jupyter
ipywidgets

compressai
foolbox
autoattack